various changes

+++ src/math/types.hpp
@@ -1,8 +1,8 @@
 #pragma once
 
-typedef float vec2[2];
-typedef float vec3[3];
-typedef float vec4[4];
+typedef double vec2[2];
+typedef double vec3[3];
+typedef double vec4[4];
 
 typedef vec2 mat2[2];
 typedef vec3 mat3[3];
